+++ vpr/src/route/rr_graph_generation/rr_graph_tile_nodes.cpp
@@ -1,5 +1,5 @@
 
-#include "rr_graph_source_sink.h"
+#include "rr_graph_tile_nodes.h"
 
 #include "rr_graph_builder.h"
 #include "rr_rc_data.h"
